refactor: move exception to domain
